Add example 1.
* WebSocketEndpointExample1.java: keeps all sessions in a map
according to the "key" they used in the connection URL. When a
session sends a message to the server endpoint, it resends the
message back to other sessions that have used the same "key".
* MessageRest.java: is a REST endpoint so anyone can send
messages to the sessions with the specific "key".
* JaxRSActivator.java: activates the JAX-RS in the application.
* example1.html: allows the user to choose a "key" to open the
WebSocket connection, to send messages to other sessions that
used the same "key", and to send messages to sessions with any
specific "key". Also, receives messages sent by other clients.